+++ plugin/src/test/kotlin/org/jetbrains/research/pynose/plugin/inspections/unittest/disabled/LackCohesionTestSmellUnittestInspectionTests.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.research.pynose.plugin.inspections.disabled
+package org.jetbrains.research.pynose.plugin.inspections.unittest.disabled
 
 import com.intellij.lang.annotation.HighlightSeverity
 import com.intellij.openapi.components.service
@@ -9,18 +9,20 @@ import org.jetbrains.research.pynose.plugin.util.AbstractTestSmellInspectionTest
 import org.junit.Test
 import org.junit.jupiter.api.BeforeAll
 
-class LackCohesionTestSmellInspectionTests : AbstractTestSmellInspectionTestWithSdk() {
+class LackCohesionTestSmellUnittestInspectionTests : AbstractTestSmellInspectionTestWithSdk() {
 
     @BeforeAll
     override fun setUp() {
         super.setUp()
         mockkObject(myFixture.project.service<TestRunnerServiceFacade>())
-        every { myFixture.project.service<TestRunnerServiceFacade>().getConfiguredTestRunner(any()) } returns "Unittests"
-        myFixture.enableInspections(LackCohesionTestSmellInspection())
+        every {
+            myFixture.project.service<TestRunnerServiceFacade>().getConfiguredTestRunner(any())
+        } returns "Unittests"
+        myFixture.enableInspections(LackCohesionTestSmellUnittestInspection())
     }
 
     override fun getTestDataPath(): String {
-        return "src/test/resources/org/jetbrains/research/pynose/plugin/inspections/data/lack_cohesion"
+        return "src/test/resources/org/jetbrains/research/pynose/plugin/inspections/data/lack_cohesion/unittest"
     }
 
     @Test
