+++ plugin/src/test/resources/org/jetbrains/research/pynose/plugin/inspections/data/lack_cohesion/pytest/test_lack_cohesion.py
