+++ plugin/src/test/resources/org/jetbrains/research/pynose/plugin/inspections/data/lack_cohesion/pytest/test_normal_cohesion.py
