+++ plugin/src/test/resources/org/jetbrains/research/pynose/plugin/inspections/data/lack_cohesion/unittest/test_lack_cohesion_no_dependency.py
@@ -1,0 +1,22 @@
+import unittest
+
+
+class SomeClass:
+
+    def __init__(self):
+        self.x = 'Hello'
+        self.y = 1
+        self.z = 5
+        self.w = "hi"
+
+    def test_inc_t(self):
+        self.y += 1
+
+    def test_dec_z(self):
+        self.z -= 1
+
+    def test_print_x(self):
+        print(self.x)
+
+    def test_check_hi(self):
+        assert self.w == "hi"
